+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog - Releases
 
 ## Unreleased
+- Add Strength indicator for JWT decryption key
 
 ## 3.4.8 (04 Nov 2022)
 - Add filter to allow the change for authentication payload